feat: rename Run type -> Runtime Type + move Runtime Version under it
- Submit form: "Run type" -> "Runtime Type", and the Runtime Version
  field now sits directly below (instead of down in Hardware & Setup).
  Reads more naturally: pick the runtime, then say which version.
- Game page filter modal + stats page dimension label: same rename so
  the vocabulary is consistent across the site.
- Runtime Version JS-injected label no longer duplicates "(Proton)" in
  the label since the type is picked in the field above.
