+++ js/profile/components/edit-modals.js
@@ -10,7 +10,7 @@ import {
   fetchReportHistory, patchUserConfig,
 } from '../api/configs.js?v=0c5650ed';
 import { updateSystem } from '../api/systems.js?v=770d14b7';
-import { notesFormattingHelpHtml } from '../../shared/submit.js?v=339c68ea';
+import { notesFormattingHelpHtml } from '../../shared/submit.js?v=bfe659f0';
 
 export let _cloudEditModal = null;
 export function getCloudEditModal() {
