+++ js/stats/charts.js
@@ -1,6 +1,6 @@
 // Stats page chart renderers.
 
-import { label, fmt, niceCeil, formatAxisLabel, vramLabel, TIER_COLORS } from './utils.js?v=c63d96b0';
+import { label, fmt, niceCeil, formatAxisLabel, vramLabel, TIER_COLORS } from './utils.js?v=2dd62302';
 import { getFilter } from './filters.js?v=f364d0eb';
 
 // Render N horizontal bar rows sorted descending by count.
